requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pyotp
SmartApi-python
